requirements.txt
django==4.1.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
django-birthday==0.1.4
django-countries==7.4.2
Pillow==9.3.0
python-telegram-bot==13.14
requests==2.26.0
matplotlib==3.6.2




